Fixed mask_range

+++ planning/autoware_diffusion_planner/include/autoware/diffusion_planner/constants.hpp
@@ -27,7 +27,7 @@ constexpr double TIMER_PERIOD_S = 0.2;
 constexpr int LOG_THROTTLE_INTERVAL_MS = 5000;
 
 // Geometric constants
-constexpr float LANE_MASK_RANGE_M = 100.0f;
+constexpr double LANE_MASK_RANGE_M = 100.0;
 constexpr double BACKWARD_PATH_LENGTH_M = 0.0;
 constexpr double FORWARD_PATH_LENGTH_M = 150.0;
